move contents of listPoints to separate component from ListPointsWrapper
